+++ app/plugins/event_manager.py
@@ -44,7 +44,8 @@ class EventManager:
             self._handlers[etype.value] = handlerList
         if handler not in handlerList:
             handlerList.append(handler)
-            log.debug(f"已注册事件：{etype.value}{handler}")
+            log.debug(f"已注册事件：{etype.value} "
+                      f"{getattr(handler, '__qualname__', handler)}")
 
     def remove_event_listener(self, etype: EventType, handler):
         """